core/connectors,doc: reposition schema param and updates documentation
The `schema` parameter of the `Records` method in an app connector is a
superset of the properties to be returned. However, the connector is not
required to ensure that the returned properties conform to this schema.
In fact, the `schema` parameter is not currently used by any connector.
Therefore, this commit moves the `schema` parameter to the last
position, as its current placement "hides" the `properties` parameter,
which actually contains the properties to be returned.

Additionally, this commit updates and improves both the internal and
public documentation of the `Records` method.