core/connectors: handle invalid last change time and continue
Previously, the sequence returned by the `App.Users` method was
interrupted without an error if the last change time of a record was
invalid (before 1900 or in the future).

This change sets the `Record.Err` field with an error and continues
processing. It also fixes the comment in the `validateLastChangeTime`
function.