assets: improve the "Action" cell of file storage connections
This commit improves the "Action" cell in the actions table of file
storage connections.

Specifically, the commit makes the action description dynamic based on
the action path and adds the file icon to give the user a way to figure
out directly from the list which file the action was created on.