core: fix cursor saving in destination app action synchronization
Previously, the cursor in a destination app action, used to synchronize
app users, was saved after each batch of synchronized users with the
last change time of the last record read. However, this approach was
incorrect, as there is no guarantee that records are returned in order
of last change time.

This change ensures that the cursor is only saved at the end of the
synchronization, using the highest last change time from the
synchronized records.