assets: limit the clickable area of the icon in the identities panel
Currently in the "Identities" panel of the user's drawer, every identity
in the list shows the name and icon of the connection from which it was
imported, which if clicked, causes the user to be redirected to the
connection page.

This commit reduces the height of the clickable area of the icon of the
connection that was previously extended to the full height of the
identity's HTML element, increasing the possibility of unintentional
clicks.