assets: disable the UI tests on the customer model page
This commit disables the remaining UI tests on the customer model page
that have not yet been disabled.

The reason for the disabling is that some of these tests keep failing
randomly from time to time.

Additionally, part of this group of tests was already disabled but since
they were meant to be run sequentially one after the other, running only
a portion of them could introduce additional problems.

We therefore proceed by disabling this group of tests completely and
then rewriting and rehabilitating them together once the root problem is
identified (see issue #1200).