connectors/dummy: fix properties handling when adding users
Previously, when a user was added without some properties, the `Records`
method did not return those properties for the newly added users, even
though they were not marked as optional required in the source schema.

This change ensures that properties present in both the source and
destination schemas are nullable. If these properties are not provided
during user creation, they will default to `null`.