assets: fix opening an action on files with no settings
This commit fixes the opening of actions on FileStorage for which the
format connector has no settings.

At the moment, in fact, the UI still tries to ask for settings,
resulting in a Bad Request. This arises from an incorrect evaluation of
'hasSettings', which is a function, not a boolean field. This commit
fixes that.