core/events/collector: fix error handling in single event encoding
This update fixes various error handling issues in the encoding of a
single event. Specifically, passing an event without properties caused
an infinite loop in the decoder.

Additionally, it ensures that in tests, if all expected events have been
decoded but an error is returned, the error is shown instead of a
generic error indicating that more events were returned than expected.
This prevents the event loop from being interrupted.

Fixes #1328