core/events, database: remove table `event_dispatching`
At the moment, Meergo's implementation is partial in relation to the
'event_dispatching' table, as it is filled but never read and/or
emptied. For this reason, we have decided at the moment to completely
remove the table (and consequently the code that manages it), and then
evaluate if and how to reintroduce this functionality later.