connectors/dummy: do proper error handling in `Upsert` method
The current error handling of the Dummy connector is wrong, because it
does not correctly implement the requirements required by Meergo, thus
falsifying the errors returned and making it appear that the export
occurred in a certain way when in fact it occurred in another.

This commit fixes the error handling, and also handles as an error the
case in which a user is exported to Dummy with an "update" when a
"create" would be needed instead (which is probably the same thing any
app would do).