assets, core: unifies support for transformation functions for events
Currently, support for transformation functions when dispatching events
to apps is inconsistent.

This commit fixes the inconsistency, allowing transformation functions
from the UI to the actual action execution.

This is the behavior that was already documented in the 'Actions.csv'
file, and on which the actions were validated (and tested).