doc: add some documentation about exporting users to apps
This documentation is intended to outline the operating principles of
exporting users to apps.

In particular, it highlights the fact that the current implementation of
export makes Meergo authoritative with respect to the data present on
the destination app (limited to specific users and properties).

In the future, we will consider expanding it to clarify it and make it
usable for a user approaching Meergo.