core: deep-validate unused action properties
Currently, on main, when validating an action only top-level properties
are checked if they are actually used. No check is done on deeper-level
properties.

Commit 0e394cc87e81b2dfae5b72c397d7e69c119ea959 made the UI send only
used properties even at a deeper level.

This commit strengthens the server-side check, so that an error is
returned even if a deeper-level property is not used.

Change and adds tests accordingly.

Resolves #1285.