assets: automatically adapt the width of the combobox list
This commit ensures that the width of the combobox list is never limited
to the width of the input to which it is paired.

Instead, the width of the widest list element is calculated during
opening, which will then be set as the width of the list itself to avoid
horizontal overflows.

Close #1338.