core,database: rename `event_write_keys.creation_time` to `created_at`
This change renames the `creation_time` column in the `event_write_keys`
table to `created_at` to align with naming conventions used elsewhere.
This change also updates the corresponding field names in the `core`
package for consistency.

Additionally, the `cursor` column’s type is changed from `timestamp(6)`
to `timestamp`, and its default value is removed as it is always
provided explicitly.