assets: remove polling from the actions list
This commit removes the logic that polls the connection in the actions
list to know if an action has been started in the background by the
scheduler.

The change thus removes the loading spinner that used to be shown on the
action's "Run now" button when an action was started by the scheduler,
as this could have caused confusion to the user who would see the
spinner potentially without understanding what triggered it.

Close #1356.