assets: display metrics numbers in a more human-readable format
This commit improves the readability of the metrics numbers: now the
numbers in the y-axis of the metrics graph are shown in an abbreviated
format (ex "K" for thousands, "M" for millions), while in the other
sections they are shown using the separator according to the format of
the "en-US" locale.

Close #1332.