assets: fix feedback not shown when executing an action
This commit fixes an issue that was occurring in the list of actions
whereby upon clicking the "Run now" button of an action no feedback was
shown, even though the action was started.

The issue occurred because, in some cases, the "ref" of the button was
recreated without being reassigned, thus preventing the handler function
from being able to call the button's methods via the ref to show
feedbacks.

Close #1367.