assets: always show the buttons to switch metrics target
This commit ensures that on the metrics page, the "Events" and "Users"
buttons used to switch the metrics target, are always shown regardless
of the type of actions currently created for that connection.

If a connection does not have (or does not support) actions on either
target, then the corresponding button will be disabled.