types: standardize fractional digits in JSON datetime
Previously, when serializing a `datetime` or `time` value to JSON, only
the most significant fractional digits were used, omitting trailing
zeros. For example, a timestamp could be represented as
"2025-03-04T14:16:44.05Z".

This change ensures that the number of fractional digits is always 0, 3,
6, or 9 while still minimizing unnecessary trailing zeros. For instance,
the same timestamp will now be formatted as "2025-03-04T14:16:44.050Z".