decimal: implement `Binary` function and method for `Decimal` encoding
This change introduces the `Binary` function that converts a
binary-encoded value into a `Decimal`. Additionally, it adds the
`Binary` method to write values in this format. The binary encoding
follows the same representation used by Parquet when storing DECIMAL
values in BYTE_ARRAY and FIXED_LEN_BYTE_ARRAY types.

Note that these functions seem to work at the moment, but only because
they pass the tests. It is not clear whether they are actually
completely correct. Also, it is likely that the implementation could be
optimized. This is discussed in #1396.

Co-authored-by: Gianluca Mondini <mondini@open2b.com>