test: remove workaround from Parquet test
There was a workaround in the test that removed the nil properties
returned by Meergo. This was not due to an API issue, as initially
seemed, but rather the way the transform function treats nil values
received from the Parquet connector.

With this now clarified, this commit modifies the transform function and
removes the workaround, simplifying the test.