connectors/filesystem: add a setting to simulate IO latency
This commit adds a setting to simulate an IO latency in addition to the
"real" latency when using the Filesystem connector.

This allows to test various components of Meergo (such as the UI) to
understand how they behave when dealing with a generic FileStorage
connector that takes some time to respond.