core: fix action failure handling and improve error reporting
Previously, if an internal error occurred during the execution of an
action, it was reported as an error in the Receive step but not as an
execution failure of the action itself. As a result, the UI incorrectly
marked the action as successfully completed when it had actually failed.

This change ensures that actions are correctly marked as failed in such
cases. Additionally, it improves error reporting for the `importUsers`
and `exportUsers` methods of *Action.

Fixes #1393