core/connectors: change handling of typed `nil` in `normalize`
This change causes the typed nil, whose type is one of the accepted
types for a certain Meergo type, to be treated as the nil of the empty
interface in the 'normalize' function.

This simplifies the behavior, effectively standardizing the cases where
a property value is any(nil) or any other type of nil.

Note that nil values implementing the 'json.Marshaler' interface are
excluded from this, see the note in the documentation added by this
commit.

The commit also adds tests and updates the documentation.