core/events/collector: fix type assertion on "context.location" property
The decoder incorrectly performed a type assertion on the value of the
"context.location" property as a `string` instead of `map[string]any`.
In fact, the type assertion can be entirely removed since what matters
is whether a value for the property is present, not the value itself.

This change fixes the issue, preventing the passed value from being
overwritten.