core/transformers/mappings: fix error message and update `parseNumber`
This change updates the error message for invalid '-' followed by a
path. Instead of "number syntax is not valid", it now returns
"the '-' character is used for negation. Did you mean the '_'
character?".

Additionally, it updates `parseNumber` to return the original string
instead of an empty string in case of error.

Fixes #1344