test: test the Parquet connector over the Apache Parquet testing data
The tests added by this commit do not include all tests. The purpose of
the commit is to add the basic infrastructure for this type of testing,
and to add some tests that already detect some unexpected behavior (see
the issues linked in the test).

Future commits will expand these tests.