cmd: add `TerminationDelay` to delay shutdown
This change adds the `TerminationDelay` setting, which specifies a
duration to wait after receiving a termination signal before initiating
shutdown. This is useful in scenarios where the process may receive a
termination signal but still needs to handle incoming HTTP requests for
a certain period.

Additionally, it logs a message when a shutdown occurs.