core: fix panic in case of event type with no schema and document it
Meergo core seems to no longer handle correctly the cases where event
types have no associated schema, perhaps it is a part that has been
little used and that does not have automated tests on it.

This commit fixes a panic that occurs in such cases and documents the
code a bit.