assets: fix error when changing the table in an existing database action
This commit fixes an error that prevented saving an existing database
action after modifying its table.

In addition, this change ensures that when the table is changed and
"Confirm" is clicked, if there are columns in the new table schema with
the same name and type as the old schema, the corresponding mappings are
preserved.

Close #1420.