assets, core: various changes to improve clarity of action events
This commit makes some visual and textual changes to improve the
usability and clarity of actions that send events.

These changes are a step toward improvement, but not necessarily
something permanent. Once we try these changes here, we will evaluate in
the future whether to keep them, continue in this direction, or remove
them.