core: return unhandled error which causes unexpected behaviors
This commit causes the transformation function creation error to be
returned instead of being ignored.

So instead of returning invalid behavior, issues like #1434 cause an
internal server error, indicating that there is another problem
elsewhere, causing the error.

This commit is therefore a step necessary to make the transformation
functions work in case the action does not yet exist and you are testing
in testing mode.

See #1434.