all: change some POST methods to GET
This commit changes some API methods so that they should be called with
the GET method instead of POST, to improve clarity on their behavior
(note that, about event retrieval, the URL path has also changed, since
by changing the method from POST to GET the conflict with the other
method that ingests events has disappeared).

This means that parameters, currently passed in the body as JSON, are
encoded in the query string.

This commit therefore changes the core, UI, tests and documentation
accordingly.

This change concerns these methods in particular (these are the new
names):

    GET /connections/:id/files/:path
    GET /connections/:id/files/:path/sheets
    GET /connections/:id/users
    GET /events
    GET /users