assets: add the "Enum" tab to the mappings comboboxes
This commit adds the "Enum" tab to the mappings comboboxes.

This new tab lists the values that the mapping expects, and in
particular indicates `true` and `false` in the case of mapping to a
property of type `boolean`, or the list of expected values in the case
of mapping to properties of type `text` that require specific values.

Close #1283.