assets: make action's ID clickable in the identities view on users page
This commit goes a step further than the current implementation, making
the action ID clickable (and aesthetically nicer and coherent with the
rest), and therefore more convenient to use the interface that shows the
list of identities for each user.

This improves usability, even just for testing.

It will perhaps be modified later to remove the ID and put the action
name, but this requires a bigger change at the implementation level, for
which issue #1445 has been opened.