assets: don't pass null/undefined `formatSettings` to APIs and document
Currently, it's a bit confusing in the UI code that certain parameters
(like 'formatSettings', in this case) are null or undefined. This commit
ensures that such values are not passed to the API, since there are
cases where this happens.

It also adds type annotations to clarify the situation a bit.