core: remove duplicated check (and thus unreachable code)
This commit removes a duplicate check. The second check was unreachable,
because if there was both a mapping and a transformation function, the
first check (in the case of mappings) would catch the case and return
the error.

By deduplicating the check, we get (1) better logical understanding,
because the check is not specific to mappings or transformations, but
both, and (2) we eliminate some unreachable, and therefore untestable,
code.

This helps #849.