core/transformers: fix `Unmarshal` by returning errors of correct type
This commit fixes the fact that 'Unmarshal' assigns a value with the
wrong type to the 'Err' field of each record, causing that type not to
be captured externally, resulting in errors being ignored.

Additionally, the new error also contains the language name, which is
useful for giving more context to the error.

Co-authored-by: Marco Gazerro <gazerro@open2b.com>