core: fix import interruption to prevent purging of unread identities
Previously, if an import was interrupted, the `Close` method of
`BatchIdentityWriter` was called, which would perform a purge if
requested. This meant that identities that were no longer being read due
to the interruption were purged from the data warehouse.

With this change, in case of an interruption, the new `Cancel` method
`BatchIdentityWriter` is used. Unlike the `Close` method, it doesn't
terminate pending operations or trigger the purge, so identities that
aren't being read are no longer deleted.

Closes #1460