all: add singular connector term for users and groups
There are parts of the UI where, for example, the term "user" exposed by
a connector needs to be singular instead of plural.

This change causes connectors to expose the singular terms in addition
to the plural terms, and these terms are then exposed through the API.

Note that this commit changes the UI to maintain the current behavior;
the use of singular instead of plural terms, where necessary, will be
done in separate commits.

The commit also changes data structures, both internal and exposed by
the API, as well as the documentation.