warehouses: store users table versions in a separate table
This change is needed to "untie" the obsolete operations table (which
will have to be removed, see issue #1475) from the still necessary code
that tracks the version of the users table.

In this way, making the two features independent, subsequent commits
will remove the old operations code, using only the new code (introduced
by PR #1473).

For #1475.