core, database: remove repaths where not necessary
Information about repaths is needed until the operations that define a
schema change are defined.

Once the operations are defined and saved in the state, there is no need
to store the repaths as well; in fact, they are saved (both in the state
and in the database) but are never actually used.

This commit, therefore, removes the repaths from all places where they
are not needed.