all: rename the "ui" to "admin"
This commit uses the term "admin" to refer to what it is currently
called "ui".

This impacts the admin URLs, as well as various expressions here and
there, documentation and tests.

It's hard to find other points to work on. If they come up, we'll change
them in later commits.

For connectors, it's right that the term "ui" remains there, because
that term generically refers to a graphical interface that the connector
exposes.