meergo: rewrite the README and introduce other files
This commit rewrites the README to orient it to a user browsing the
repository to discover Meergo, rather than to orient it to us for
internal software development.

It also adds two files CONTRIBUTING.md and DEVELOPERS.md, the latter of
which contains the information that was previously in README.

The texts now in the README.md are just a draft and clearly need to be
reviewed and extended.