assets/tests: do not include trailing slash in variable `adminURL`
The inclusion of the slash at the end of 'adminURL' is confusing,
because then when the variable is used it is not clear whether and how
it should be separated from the other components of the path.

This commit therefore removes the slash from the variable and adds it
where necessary, in a way similar to how it is done in other contexts,
such as Unix shells.