cmd/meergo, docker-compose: reformat YAML files
This commit ensures that all YAML files in the repository are formatted
according to the RedHat YAML formatter.

If in the future we find another better or more convenient formatter we
can use that one, otherwise for now we use this one, to avoid "bouncing"
the formatting changes to the YAML files between one PC and another.