cmd: fix external URL shown when Meergo is started
The fact that the eventual trailing slash on the left instead of the
right was (erroneously) removed meant that if the external URL contained
a trailing slash in the configuration file, what was then displayed
contained a double slash, and therefore the link was broken.

This commit fixes the behavior.