all: add configuration for the CDN URL and the event URL
This commit adds two settings:

* the URL of the CDN that serves Meergo files, including admin and
  JavaScript SDK
* the URL of the endpoint that receives events

These settings are managed in the configuration file, and then exposed
to the admin through two methods of the API dedicated to the admin, and
therefore not documented.

This change allows to correct the URLs shown in the snippet of the
connections that import events.

Issue #1502 tracks the possibility of moving these settings from the
configuration file to the organization level; however, this is not
essential for a first version of the software, so we have decided to
postpone this implementation and limit ourselves for now to the
configuration YAML file.

This commit also adds tests.