assets: hide the password error at every password update
This commit ensures that on the password reset page, any validation
error associated with the currently entered password is hidden each time
the password is updated.

This is necessary to ensure that in the event that the updated password
successfully passes validation, the now obsolete error message is not
displayed.