all: configure Meergo through environment variables instead of YAML file
This commit makes Meergo configuration happen via environment variables
instead of the 'config.yaml' file.

This is especially convenient, for example, in cases where Meergo is
started via Docker Compose.

Also, related to this, it adds support for the '.env' file, which is
sourced when Meergo is started.

It updates everything related to this, including the testing
infrastructure, starting via Docker, and documentation.

While it doesn't exactly implement the solution proposed in the issue,
this commit closes #1063.

Closes #1063.