all: send telemetry to Sentry (from the server)
This commit causes Meergo to send telemetry to Sentry.

This change is only server related. A change will be made to the admin
later.

Some naming logically conflicts with OpenTelemetry are tracked by
the issue #1516 and will be fixed later.

The revision of the type of information to send and the fact that user
data is not sent is tracked by issue #1517.

Telemetry to Sentry is enabled by default. To disable it, set this
environment variable:

    MEERGO_DISABLE_TELEMETRY="true"

This commit also modifies Docker Compose and Meergo tests.