cmd, doc: make the OpenTelemetry env variable more specific
There is currently a logical conflict between the telemetry sent to
Sentry (which is what is currently working, sending errors) and the
telemetry sent to the OpenTelemetry collector, which is minimal and
experimental.

This commit causes the environment variable 'MEERGO_TELEMETRY_ENABLE',
which controls sending telemetry to OpenTelemetry, to be renamed to
'MEERGO_OPEN_TELEMETRY_ENABLE' and documented more specifically,
explicitly stating that this is an experimental feature and that the
actual telemetry is handled by the other variable,
'MEERGO_DISABLE_TELEMETRY'.

Other commits may clarify this further.

For #1516.