all: rename the 'telemetry' package to 'opentelemetry'
There is currently a logical conflict between the telemetry sent to
Sentry (which is what is currently working, sending errors) and the
telemetry sent to the OpenTelemetry collector, which is minimal and
experimental.

This commit reinforces the difference by renaming the package that sends
data to the OpenTelemetry collector, to avoid the logical conflict.

For #1516.