test: remove workarounds for #1172
After doing some tests on Windows, @retini and I have found that the
problem reported by issue #1172 is no longer reproducible.

In the tests, since we use testcontainers for databases, there is no
longer the problem of resolving the name "localhost"; while for what
concerns the execution of Meergo itself, even there the name "localhost"
seems to be resolved correctly even on Windows.

Perhaps this was caused by some issue resolved over time, perhaps an
updated package or a new version of the Go runtime.

For this reason, issue #1172 is closed.

Closes #1172.

Co-authored-by: Andrea Retini <retini@open2b.com>