all: add "from" configuration for member emails
Currently, emails sent to members have the "From" field set to the SMTP
server username. This is incorrect.

This commit therefore introduces a new environment variable
'MEERGO_MEMBER_EMAIL_FROM' to configure the "From" field.

It is mandatory to provide a value for this variable when sending emails
to members, otherwise the emails cannot be sent.