meergo: rearrange and improve doc. of Docker Compose file
This commit rearranges the 'compose.yaml' file and adds documentation to
it, so that the user trying out Meergo can better understand how to
configure it and intervene in that configuration, if necessary.

It also adds references to how to set up another PostgreSQL instance
(e.g. one running locally) in place of the one started with Docker
Compose.

For warehouses, the discussion is similar to the one documented. The
host 'host.docker.internal' can be used to refer to the host network
where the warehouse runs.

Resolves #1522.