core/state: add `connector` parameter to `Workspace.AccountByCode`
Previously, the `Workspace.AccountByCode` method returned an account
from the workspace based on its code. However, since it didn't allow
specifying the connector, it could return the account from any
connector.

This change adds a `connector` parameter to the method, allowing the
caller to specify which connector's account to retrieve.