.github/workflows: add skeleton for the Action that uploads source maps
This commit adds the skeleton file of the GitHub Action that sends
source maps of admin files to Sentry.

This change is being implemented in PR #1521, but it is necessary to add
this file to 'main' to make GitHub give the ability to run an Action
that is on another branch. As far as I know, there is no other way than
this to develop a GitHub Action that is on a branch other than 'main'.