assets: integrate Sentry for error monitoring
This commit integrates Sentry into the admin for error monitoring.

This also involves a change to the part in Go that generates the assets,
so that the Debug IDs used by Sentry for the correct association of
errors with the respective source maps are injected.

Furthermore, it adds a GitHub Action to upload to Sentry the necessary
information, indicated in the previous point.

Resolves #1518.

Co-authored-by: Gianluca Mondini <mondini@open2b.com>