.github/workflows: try increasing the tests timeout to 30 minutes
Currently the GitHub Action that runs the tests fails because the 20
minute timeout expires while the tests are still running.

This commit therefore increases the timeout to 30 minutes, in the idea
that this is enough to finish the tests, and with a view to slowly
investigate and speed up the tests, which currently take a really long
time to run.