assets/test: use `toBeAttached()` instead of `toBeVisible()`
This commit ensures that the UI tests always use the `toBeAttached()`
method instead of `toBeVisible()` to check that certain elements are
present in the UI for testing purposes.

This change is necessary to prevent the tests from failing randomly just
because an overlay element (e.g., a tooltip, a dialog) hides the portion
of the UI that is being checked with `toBeVisible()`, when in real use
this would not be a problem because the user could simply scroll to make
that same portion of the UI visible.

The use of `toBeAttached()` therefore turns out to be enough to test in
our use case, ensuring that a given element is or is not present in the
UI regardless of its temporary visibility.