assets,cmd: consistently use `LegalCommentsEndOfFile` in esbuild
This change fixes an inconsistency in the use of esbuild's
`LegalComments` option by always setting it to
`LegalCommentsEndOfFile`, ensuring legal comments are appended at the
end of the bundle.

When running the vendor process, this option’s value is irrelevant
since the bundle is generated solely to identify dependencies for the
vendor, and this setting has no effect in that context.

Fixes #1527