all: add an installation ID to the database and send it to Sentry
This commit:

(1) adds an installation identifier to the PostgreSQL database, a UUID
that is generated when the Meergo database is initialized;

(2) sends the installation ID to Sentry along with each event, so that
Sentry can filter and categorize events based on the Meergo installation.

Note that at this time, the installation ID is only sent to Sentry for
events sent from the server. For the admin, it is not sent at this time.