all: add overview documentation to connectors
Connectors previously only included a short summary description. This 
change introduces a new "overview" section for each connector, which 
can vary depending on the source and destination.

At this stage, the overviews are added but left empty. Future commits 
will populate them and make them visible in the admin interface.