assets: show the connector's documentation
This commit ensures that when a connector is clicked from the list of
connectors, instead of going directly to the configuration page, a
drawer containing the connector's documentation is opened.

From the drawer it is then possible to click a button to go to the
configuration page and add the connection.