connectors, doc: implement the Meergo API connector
The Meergo API connector is a connector that allows to interface to the
Meergo APIs from any language or application, as long as it can make
HTTP requests.

At the moment, when a user tries to import events and users from a
language/application for which there was no SDK, it is necessary to add
a connector for another language (for which an SDK is available), just
to be able to call the APIs.

This commit implements the Meergo API connector to address this use
case, and it also adds the necessary documentation.