assets: fix ordering of categories filter in connectors adding page
After the "API" connector category was added, the "All" pseudo-category
moved to second place, as categories are sorted alphabetically in the
admin.

This commit therefore first sorts the categories alphabetically, and
then adds the "All" pseudo-category as a test, so that it is always
first, regardless of the names of other categories.