all: normalize file endings in Markdown and SQL files
This commit causes all Markdown and SQL files (excluding vendored ones)
that we have in the repository to always end with a single line break.

So, in fact, this commit removes all whitespaces (spaces, newlines,
tabs, etc...) at the end of Markdown and SQL files, then append at the
end of that files just a single newline character.

This, in addition to logically standardizing newlines, avoids the
annoying warning about newline changes, or the absence of newlines,
reported in Git and GitHub diffs.