test/meergotester: refactor the way that Meergo is built for the tests
Currently, when Meergo is compiled and run as an external process to
launch tests (which is the default case), the compilation takes place
inside the local repository directory 'cmd/meergo', which may contain
other user files and may not include the necessary connectors for
running the tests.

This is obviously chaotic and incorrect.

This commit therefore rewrites the way Meergo is compiled for tests,
creating a temporary directory and executing within it the procedure
that a user would follow to compile Meergo using their chosen
connectors, with the only difference being that in this case, the source
of Meergo used to compile the executable is taken from the local
repository, not from the remote one.