.github/workflows: try increasing the tests timeout to 1 hour
Currently the GitHub Action that runs the tests fails because the 30
minutes timeout expires while the tests are still running.

This commit therefore increases the timeout to 60 minutes, in the idea
that this is enough to finish the tests, and with a view to slowly
investigate and speed up the tests, which currently take a really long
time to run.

Note that we had already made a change in the past to increase the test
timeout from 20 minutes to 30 minutes: https://github.com/open2b/meergo/commit/f817d25780206e0bbb7e7c2a074b63ea589856cd.

For #1531.