meergo, cmd: detach Sentry from slog, sending only panics from server
This commit causes the Meergo server to no longer send errors to Sentry
via slog, but only send panics, at least for now.

This is because it is not possible to send the errors returned by the
various methods and functions using the Sentry integration with 'slog',
as they may contain personal and/or sensitive data. As for the panics,
we ensure that they never contain personal and/or sensitive data.

For #1517.

Resolves #1538.