doc: lock the sidebar during scrolling
This commit ensures that when scrolling the page, the sidebar remains
fixed at the top, so that the user can always see the list of
documentation pages and the currently selected page.

In case the sidebar is higher than the viewport it will be automatically
made scrollable, so that the entire list can be seen.